requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
aiosqlite==0.20.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
boto3==1.35.76
botocore==1.35.76
httpx==0.28.1
python-dateutil==2.9.0
pytz==2024.2
email-validator==2.2.0
bcrypt==4.2.1
cryptography==43.0.3
pyjwt==2.10.1
requests==2.32.3
aiofiles==24.1.0
starlette==0.41.3
anyio==4.7.0
typing-extensions==4.12.2
greenlet==3.1.1
cffi==1.17.1
pycparser==2.22
certifi==2024.12.14
charset-normalizer==3.4.0
idna==3.10
urllib3==2.2.3
sniffio==1.3.1
h11==0.14.0
click==8.1.8
six==1.17.0